Add HitManager class for hit logic
